refactor: question/domain/entity에 있는 엔티티를 global/entity로 이동
