+++ src/main/java/com/upstage/devup/global/entity/Answer.java
@@ -1,6 +1,5 @@
 package com.upstage.devup.global.entity;
 
-import com.upstage.devup.question.domain.entity.Question;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
