+++ src/main/java/com/upstage/devup/user/answer/service/UserAnswerSaveService.java
@@ -3,7 +3,7 @@ package com.upstage.devup.user.answer.service;
 import com.upstage.devup.global.entity.User;
 import com.upstage.devup.global.exception.EntityNotFoundException;
 import com.upstage.devup.question.domain.dto.QuestionDetailDto;
-import com.upstage.devup.question.domain.entity.Question;
+import com.upstage.devup.global.entity.Question;
 import com.upstage.devup.question.service.BoardService;
 import com.upstage.devup.user.answer.domain.dto.UserAnswerDetailDto;
 import com.upstage.devup.user.answer.domain.dto.UserAnswerSaveRequest;
