+++ src/main/java/com/upstage/devup/user/statistics/service/UserWrongAnswerReadService.java
@@ -1,7 +1,7 @@
 package com.upstage.devup.user.statistics.service;
 
 import com.upstage.devup.global.exception.EntityNotFoundException;
-import com.upstage.devup.question.domain.entity.Question;
+import com.upstage.devup.global.entity.Question;
 import com.upstage.devup.user.answer.domain.entity.UserWrongAnswer;
 import com.upstage.devup.user.answer.repository.UserWrongAnswerRepository;
 import com.upstage.devup.user.statistics.domain.dto.WrongNoteSummaryDto;
